- Added Jim's Makefile.linux; - Changed: Decal scales now use full precision fixed point numbers. - Changed: Keeping impact decals in their own statlist is enough to keep track   of them for when one needs to be destroyed. There's no need to maintain a   separate list for them. - Fixed: Decal actors did not spread their decals across neighboring walls. - Fixed: Decal groups did not initialize their IDs and could not be reliably   used with the decal actor. - Fixed: Decals on moving polyobjects were not interpolated. R_RenderDecal()   now uses the decal's LeftDistance to calculate its visible location, so it   always stays in sync with the wall's vertices. This also lets me dump some   code from the polyobjects that maintained the decals' (x, y) coordinates.   Also, the decals' x and y information is redundant and can be removed.   Doing this revealed a bug with slider decals and horizontal sliding:   That is, it didn't work at all. I have opted to simply remove the horizontal   sliding support so that I don't have to worry about what happens when a   decal slides across the edge of a wall. - Fixed: DBaseDecal::LeftDistance was calculated as a 30.2 fixed point number.   It should be 2.30 fixed point.
SVN r35 (trunk)

+++ src/g_shared/a_keys.cpp
@@ -28,7 +28,7 @@ struct Keygroup
 
 	bool check(AActor * owner)
 	{
-		for(size_t i=0;i<anykeylist.Size();i++)
+		for(unsigned int i=0;i<anykeylist.Size();i++)
 		{
 			if (anykeylist[i].check(owner)) return true;
 		}
@@ -52,7 +52,7 @@ struct Lock
 
 	~Lock()
 	{
-		for(size_t i=0;i<keylist.Size();i++) delete keylist[i];
+		for(unsigned int i=0;i<keylist.Size();i++) delete keylist[i];
 		keylist.Clear();
 		if (message) delete [] message;
 		if (remotemsg) delete [] remotemsg;
@@ -71,7 +71,7 @@ struct Lock
 				}
 			}
 		}
-		else for(size_t i=0;i<keylist.Size();i++)
+		else for(unsigned int i=0;i<keylist.Size();i++)
 		{
 			if (!keylist[i]->check(owner)) return false;
 		}
